Preliminary changes needed for orchestration to work with MEASURE xml source and VALIDATION.
